requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
langchain-openai
python-dotenv
pydantic
